lxd/storage_volumes_state: Return actual volume size for stopped intances
When getting instance disk total size of a stopped instance, you want to see what size is the root disk you will be dealing with when you start it. But it could be the case that the current volume size is different from the instance configuration as the volume when gets resizes (if needed) when the instance starts. So for the volume state, we always get the actual current size of the volume.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>