lxd/devices: Rework `ueventParseVendorProduct` logic
This commit reworks the logic in `ueventParseVendorProduct` to allow
adding `unix-hotplug` devices with only `subsystem` specified. Rather
than an early return for devices with subsystems that don't match
`hidraw`, we gather the missing device info (`vendorid`/`productid`)
when device system is `hidraw`. Otherwise, the `vendorid` and
`productid` found in the config are returned.

Signed-off-by: Kadin Sayani <kadin.sayani@canonical.com>