lxd: Remove lint exception for defer rule.
The defer rule was removed from our golangci-lint configuration in
5f926194457f12cfe356f2d3bc025465b04ab513. With this change, the
defer in this function (which is called in a for loop) no longer needs
to be ignored by the linter.

The defer rule was originally removed because refactoring existing code
to stop calling defer within loops was resulting in subtle bugs that
were hard to reason about.

We have extensive tests for LXD, so we should:
- Avoid calling defer inside a for loop, and watch for these in pull
  requests.
- Focus on improving code quality from a TICS perspective first.

Signed-off-by: Mark Laing <mark.laing@canonical.com>