shared/cmd: Update RenderSlice for SQL format.
This function contained a switch statement of table formats. It wasn't
obvious that this needed to be updated when I added the SQL format. This
commit adds a function `isTableFormat` which returns true if the format
is actually a table and not JSON or YAML. `RenderSlice` is updated to
use that function so that data transformations are only applied when
necessary.

Signed-off-by: Mark Laing <mark.laing@canonical.com>