lxd/patches: Remove volatile.%.last_state.ip_addresses keys more efficiently
The patchInstanceRemoveVolatileLastStateIPAddresses previously loaded all instances
and snapshots, then matched the specific keys to remove and removed each one by a
separate query.

Although this allows better debug logging, it was causing problems with users who
have a lot of instances or snapshors from starting LXD as the patch was taking over
10s to apply and the single transaction was timing out.

This commit fixes the issue by pushing the matching logic down into the query
and removing the relevant rows in a single query per table.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>