lxd/resources/cpu: fix GetCPU() to handle multisockets systems
When VM has more than 9 vCPUs and they are exposed as sitting
on a different "physical" sockets we don't handle it properly [1]
as we implicitly assume that entries in /sys/devices/system/cpu/ are
sorted the same way as in /proc/cpuinfo. But it's not correct assumption.

To reproduce that, the following QEMU command line parameter can be used:
-smp 12,sockets=12,cores=1,threads=1

Affected systems have more than one line in the following command output:
cat /sys/devices/system/cpu/cpu*/topology/physical_package_id | uniq
and more than 9 CPUs.

[1] https://github.com/canonical/lxd/issues/14159#issuecomment-2551838382
Fixes: #14159
Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>