lxd: Add reverter to storagePoolVolumeUpdateUsers
storagePoolVolumeUpdateUsers is used in two places, one of which does not
currently use a reverter. If that update fails, the instances/profiles
could be left in an inconsistent state.

This pushes the reverter inside the function, so that it is safe to call
without needing a reverter.

My hunch is that this is also slightly less slow in case of failure.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>
(cherry picked from commit 2064793928e0532c62afa7f62f5d470909adc8dd)