lxd-agent: If the target path of a mount is relative, make it absolute.
If the path is not absolute, the mount will be created at `/run/lxd_agent/<path>`,
since the mount command executed below originates from the `lxd-agent` binary that is in the `/run/lxd_agent` directory.
This is not ideal and not consistent with the way mounts are handled with containers. We then make the path absolute.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>
(cherry picked from commit a263d0e31994ad71e9ee929941edb5b8e8762c1e)