lxd/storage/drivers/powerflex: Lock the API token
This ensures a token which is valid isn't reset by another caller.
Also return the valid token so it cannot be modified anymore which effectively
detaches it from the internal struct which is protected by the lock.

It didn't cause any issues during tests but improves the token handling in general.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>