shared/util: Introduce the `EncodeRemoteAbsPathWithNonExistingDir` function
When creating a directory within an instance filesystem, we should be able to track
the directory structure that has been created in order to remove the chain of directories
when removing a disk device associated to an instance.

This function takes an sftp client that, given an absolute path (this path is the desired directory path that a user whishes to create in the instance),
it creates a path encoding like the following:

<existing-path-part>/./<non-existing-path-part>
                   ~~~~
	          SEP_MARK

Using this encoding, when unmounting a device associated to a created directory within an instance, we can remove the chain of created directories, now that we know which have been created and which were there before.
In the end, we remove at the path `<existing-path-part>/<non-existing-path-part>[0]` (for the VM case)

For the container case, we remove directories using the sftp client and since this one doesn't have a recursive remove API for directories,
we need to directories from the deepest level all the way up:

1) rm `<existing-path-part>/<non-existing-path-part>[:L]`
2) rm `<existing-path-part>/<non-existing-path-part>[:L-1]`
3) rm `<existing-path-part>/<non-existing-path-part>[:L-2]`
...

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>
(cherry picked from commit f13ac979c805bd9417cf78be0e83a2c97b9382a1)
Signed-off-by: Nikita Mezhenskyi <nikita.mezhenskyi@canonical.com>