lxd/storage/utils: Populate `volatile.rootfs.size` on volume creation & delete effective fields
This contains a key part of the logic of keeping track of volume sizes
on the database. This commit stablishes that `VolumeDBCreate` should
receive the volume config as it will be used for volume creation,
including its intended size. Then this config would be copied (as to not
mess up the original config passed in) and adapted to be suitable for
going into the database.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>