lxd/storage_volumes: Fix copying storage volumes between cluster nodes when target and project parameters are set
This commit includes the request project name in
`clusterCopyCustomVolumeInternal` when building the `storageVolumePost`
request; this is required to ensure the request project name is not
always default when copying storage volumes with a target project set.
This commit also fixes storage volume migration error handling by
checking if effective project and target project are the same only when
the target project is not default.

Signed-off-by: Kadin Sayani <kadin.sayani@canonical.com>