lxd/storage/drivers/powerflex: Ensure mapVolume returns safe reverter
Previously the mapVolume returned a reverter that disconnected NVMe session
in case of an error. Since mapVolume acquires a lock, calling disconnect
outside of the mapVolume or unmapVolume functions may result in a race
condition because another operation could establish a connection before the
reverter is called. In such case, the reverter would carelessly just terminate
the newly estabished connection.

Therefore, the outerReverter ensures that instead of disconnecting from the volume
the unmapVolume is called, which acquires the lock and ensures the disconnect is
called only if the session is not in used.

Co-authored-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
Co-authored-by: Din Music <din.music@canonical.com>
Signed-off-by: Din Music <din.music@canonical.com>