lxd/device: Clarify "pool" and "volume" in Disk*RBDFormat
Different calls of these functions are currently using both senses of volume:
- A Ceph block device (RBD Image)
- A LXD storage volume

Because a disk device can specify an RBD image as "ceph:<pool_name>/<volume_name>",
we'll go for the first meaning; the name passed to these functions should
be the name of a RADOS block device (RBD image), not a LXD storage volume.

These functions should (in future) be replaced with a more comprehensive
abstraction for what "MountInfo" is currently doing. This serialization
is not written anywhere, it's just deserialized when it gets to the
instance driver.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>
(cherry picked from commit 0575f2618af9b8ef15ae3e2435aba39d6d153881)