shared/api: Explicitly report null usage
This omitempty is ineffective here since 0 is not a valid value for usage and any error in retrieving usage resulted, up until this point, in a null `StorageVolumeStateUsage` struct. So changing this doesn't break any previously existing behavior and this now matches how we represent unsupported DISK usage, by showing 0 usage.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>