lxd/loki: Rework use of context
This commit removes a context stored in the `Client` struct in favor of
passing context as the first function argument as needed. A default
timeout is set on a short lived context for use in `send`, and
`sendBatch` will now stop retrying when the `quit` channel is closed by
the `Stop` function.

Signed-off-by: Kadin Sayani <kadin.sayani@canonical.com>