client: Update client to use additional scopes.
This changes the following:
- Stops using the value of `X-LXD-OIDC-groups-claim` as a scope.
- Stops requesting `offline_access` and `profile` by default.
- Gets additional scopes by unmarshalling the value of
  `X-LXD-OIDC-additional-scopes`.

If an old client connects to a new LXD server, they will continue to
request `offline_access` and `profile` scopes by default, but will not
request any additional scopes. Nor will they receive the groups claim
header.

If a new client connects to an old server, they will only request the
`openid` and `email` scopes. This is sufficient for logging in in most
cases.

Signed-off-by: Mark Laing <mark.laing@canonical.com>