client: Update client to use `X-LXD-OIDC-scopes` header.
Updates the `DoHTTP` function on `ProtocolLXD` to check for the
`oidc_scopes` API extension if the `oidcClient` is non-nil. The presence
of this API extension is passed to `(oidcClient).do` to modify
authentication behaviour based on the expected headers.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 132454baf6be255cbf415fa52a4a8b71f83c1042)