lxd/storage/connectors/utils: Allow setting custom timeout for connect function
If context has not deadline configured, the default 30 second timeout is applied.
However, if context has deadline configured, it is still wrapped into another context
to ensure we can manually cancel the ongoing context. This is required when function
exits on first sucessfully established connection. Other connection attempts remain
running in the background, and we want to cancel them only if the returned reverter
is called.

Signed-off-by: Din Music <din.music@canonical.com>