lxd/storage/utils: Don't fill config on `VolumeDBCreate`
Relying on VolumeDBCreate to perform this adds unnecessary contraints.
Such as not enabling to perform operations that require having a filled in volume
config before create the DB volume unless one adds a redundant call to FillVolumeConfig.
Or not being able to set keys that aren't supposed to go in the database before calling
VolumeDBCreate, such as `size` or `state.size`.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>