lxd/storage/utils: Don't fill up config on `VolumeDBCreate`
`VolumeDBCreate` has a weird functionality of filling the volume with default config for its pool. Since the config keys that are filled up config affect the original config passed in and are sometimes necessary for subsequent operations on the volume beyond creating the DB entry for the volume. So this is part of the plan to move this operation elsewhere.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>