lxd/db/cluster: Check group names on membership change failure.
This function was implemented with the idea of keeping the happy path as
fast as possible. For that reason, the group names aren't checked and
instead we rely on the database call to do that for us. If we didn't
write the expected number of rows, the likely cause is that one of the
groups was not defined. At this point we're no longer on the happy path,
so check which group names were not present and return an error.

Signed-off-by: Mark Laing <mark.laing@canonical.com>