lxd: Update the `waitForOperations` function to take a map of entity URLs to a usage bool.
The waitForOperations should be executed concurrently in order to avoid a long running operation to block
the daemon Stop function. However, we must ensure that we're not shutting down entities like instances or storage volumes
further down in the daemon Stop function, that might have an associated operation.

We communicate this state of entities being used or not in the `resourceMap` (with a mutex for concurrent access)
If `false` is tied to an entity URL (or just that there is no entry) it means that an entity is safe for being shutdown/unmounted. Else, we must wait.

In the daemon side, we also added a `firstOperationsCheckPass` channel that wait for a first scan of the operations before calling the other goroutines so that we have an initial state map to work with.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>