lxd: update the `instancesShutdown` function to work with a usedResourcesMap
Now the `instancesShutdown` is being able to query the state of instances (busy or not).
If the instance to be shutdown in a worker is detected as busy, we send it back to the shutdown channel with a bounded exponential backoff.

Backoff strategy:

It starts with an initial retry delay of 5 seconds.  Subsequent retries increase exponentially, multiplying the previous delay by 1.5, with a jitter factor of 0.2 to avoid synchronized retries. The delay is capped at a maximum of 60 seconds to prevent excessively long waits. The entire retry process continues for a maximum of 15 minutes (MaxElapsedTime), after which it terminates regardless of success, ensuring the task doesn't run indefinitely.

Example:

Retry 1: 5 seconds
Retry 2: 7.5 seconds (5 * 1.5)
Retry 3: 11.25 seconds (7.5 * 1.5)
Retry 4: 16.875 seconds (11.25 * 1.5)
Retry 5: 25.3125 seconds (16.875 * 1.5)
Retry 6: 37.96875 seconds (25.312 * 1.5)
Retry 7: 56.953 seconds
Retry 8: 60 seconds, because MaxInterval will prevent from incrementing more so we do 60s from here.
Retry 9: 60s
Retry 10: 60s
Retry 11: 60s

<Instance operation finishes before the 15min max elapsed time!>

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>