lxd/db/cluster: Add fine-grained TLS identities to ID from URL query.
The database queries for certificate and identity entity types were
changed in #14173 so that they don't overlap. This included restricting
the ID from URL query for identities to only include OIDC identities
(which were the only fine-grained identities at the time).

When fine-grained TLS identities were added in #14207, the new identity
types were not added to the ID from URL query. This prevents permission
subsystems from resolving the URL of a pending or activated fine-grained
TLS identity, and therefore prevents permissions from being granted against
identities of this type.

Signed-off-by: Mark Laing <mark.laing@canonical.com>