lxd/operations: Introduce the 'OperationTracker' struct
The OperationTracker and the tracking groups contained in it, is a handy way to keep track
of operation objects by associating them with their entity URLs. This operation tracker
is meant to track instances and volumes related entities in a separate fashion so that
we don't hold a global mutex to take a look to an entity if we want to look an an other entity of a different type at the same type.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>