lxd/operations: Replace `waitForOperations` by `entityToPendingOperations`
`waitForOperations` is only used in the LXD shutdown sequence Stop() function at it is a blocking function.
We replace it by a non-blocking function that simply populate an OperationTracker object that will be used later.

It is also worth mentionning that once the LXD shutdown sequence is triggered, no more operation can be added,
so I don't see the use of "waiting".

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>