lxd/db/cluster: Permissions must be deleted for auth groups.
This was an oversight. The removal of permissions associated
with a group was not performed because auth groups are referenced
by permissions by foreign key. When an auth group is deleted, all
permissions for that auth group are also deleted. However, when an
auth group is granted permissions against another auth group that
is subsequently deleted, those permissions are referenced *only by
their entity type and ID*, and therefore must be deleted via trigger.

Signed-off-by: Mark Laing <mark.laing@canonical.com>