lxd/db/cluster: Delete certificate permissions on identity deletion.
Similar to #15066, the splitting of certificate and identity entity
type definitions complicates the deletion trigger for the identities
table. We need to delete permissions for both certificate and identity
types when an identity is deleted.

Note that since certificates and fine-grained identities are stored in
the same table, it is not possible for this deletion trigger to reference
an incorrect permission or warning.

Signed-off-by: Mark Laing <mark.laing@canonical.com>