lxd/instance/instancetype: Introduce 'volatile.cpu_flags' config key for VM
This key gets populated when a VM is exported. This config key will then be stored in the backup index file
and will be available when imported in a new lxd node.

This is useful when we want to import a stateful VM across different LXD nodes that are not in a cluster (if we're not in a cluster, we can't get the CPU flags through the /1.0/resources endpoint of the source node so we have to embed the info as a volatile key)

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>