lxd/operations: Replace waitForOperations by `pendingInstanceOperations`
`pendingInstanceOperations` is only used in the LXD shutdown sequence Stop() function at it is a blocking function.
We replace it by a non-blocking function that simply populate a map of instance URL to their operation.

It is also worth mentionning that once the LXD shutdown sequence is triggered, no more operation can be added,
so I don't see the use of "waiting".

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>
(cherry picked from commit 987aa37f386821c594115d05e7151d5524ef8072)