lxd/daemon_storage: refactor `daemonStorageVolumesUnmount`
We will no longer call `daemonStorageVolumesUnmount` in a separate goroutine but after the instancesShutdown function.
However, we still need to be able to cancel the volume unmount after a timeout, hence the introduction of the 'select' statement with a context handling.

This change also enable a context to cancel an ongoing unmount procedure.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>