snapcraft/qemu: Add QEMU patches for better error handling
This should bring more visibility about what is failing when an error occurs
in the `kvm_arch_get/put_registers` function calls.

These issues might arise when migrating a VM from CPU A to similar CPU B (even if we keep a common set of CPU flags between CPU A and B)
Now, we should be able to see what the exact issue is.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>