lxd/bgp: Rework start/stop logic
This simplifies the BGP package a bit and improves on the ability to
reconfigure it.

The go-bgp ability to start/stop the daemon doesn't work reliably, so we
need to spawn a completely new listener instead.

This moves Start/Stop interactions to the main Reconfigure option and
makes sure to always feed all peers and prefixes on startup.

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
(cherry picked from commit 6a0d0137735c260b5872e27c7d94720e8d989228)
Signed-off-by: Nikita Mezhenskyi <nikita.mezhenskyi@canonical.com>
License: Apache-2.0