lxd/device/disk: Fix attachment of ceph instance volumes in default project
The startVM function was always using the project.StorageVolume() function to derive
the underlying ceph rbd image name. This is incorrect when the volume being attached
is from an instance in the default project because the `default_` prefix is only added
for custom volumes, not instance volumes.

Additionally this function was also using the instance or volume's project name directly
rather than using the existing `storageProjectName := project.StorageVolumeProjectFromRecord()`
variable. This is a problem because for custom volumes it wasn't taking into account the
value of `features.storage.volumes` set on the project (indicating the effective project
for a custom volume).

So instead I switched to using the same invocation pattern we already have from
disk.mountPoolVolume() function.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>