lxc: Update `lxc config device` command with refactored completions.
If the device type is "nic" or "gpu", guide the user with completions
to `nictype`, `gputype`, or `network` so that the subtype is likely to
be set when getting all device options. This will narrow the result set,
making valid options more discoverable.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit f7eff096c700625afe96dab61c9554ff49f56fe6)