lxc: Refactor file completion.
`cmpFiles` was used to append a `/` to instance names and show
local files if requested. Some lxc commands (e.g. import) need to
compare local files. Unfortunately, since the first argument to
import is optionally a remote name, we can't return the default
file comparison directive, so we need a go method only for
comparing local files.

This is good though, because it means we can also filter by file
extension to show more relevant results.

Signed-off-by: Mark Laing <mark.laing@canonical.com>