client/events: Introduce EventListenerManager
Introduction of the eventListenerManager aims to solve two issues, the parent reference and code duplication.

The EventListener has no longer a reference to the parent (ProtocolLXD), which allows us to reuse the
EventListener for ProtocolDevLXD as well. Another benefit is that the setup and tracking of event listeners
is now unified for both ProtocolLXD and ProtocolDevLXD, reducing code duplication.

Signed-off-by: Din Music <din.music@canonical.com>