lxd: Update the users in storagePoolVolumeTypePostRename after the actual volume rename
This ensures that the volume including its database entry are already renamed before updating the volume users.
As part of updating the users, the changes are getting persisted to file for every instance.
As the files contents are loaded from the database, the entries already need to be up to date.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>