shared/entity: Only set the `target` parameter if location is required.
When moving an instance, the client calls
`POST /1.0/instances/{name}?target={member}`, where `member` is the
cluster member that the instance is being moved to - not it's current
location.

When the auth subsystem encounters this URL, it calls
`(lxd/db/cluster).GetEntityReferenceFromURL` which parses the URL and
passes the project, location, and path arguments into an entity type
specific query. The query for instances expects the `location` to be
empty, because instances are not cluster member specific.

We can't update the query for instances to use their current location
(as they are not cluster member specific), so instead this commit
updates the `(shared/entity.Type).URL` function to only add the `target`
parameter if it is required by the entity type.

This works because all URLs are normalised by the OpenFGA driver to
e.g. add the project parameter if not present. Therefore, the location
of all entity types are not cluster member specific is now no longer
included in any authorization checks.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 0605c1e78a97b506337e3c489e6726ce41d208af)