test/suites: Update clustering move tests to use fine-grained auth.
This commit updates the clustering move suite to use a fine-grained TLS
identity for move operations. It creates a group with the minimum
permissions required to perform the same assertions as before this
change. The permissions are per-instance so that we can be sure
permissions are persisted after a move.

Additionally, this test now asserts that the
`can_override_cluster_target_restriction` entitlement on `server` is
working as intended.

Signed-off-by: Mark Laing <mark.laing@canonical.com>