Makefile: add `liblxc` target
This builds a very minimal `liblxc` that intentionally does not include the
`lxc-*` tools nor support for bits that are unneeded from a LXD perspective.

This speeds up the compilation too (need to compile 62 units instead of ~550).
It also reduces the size of the `liblxc.so` file and the amount of shared libs
it depends on.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>