lxd: Disallow deletion of server certificates.
It makes sense to be able to update a server certificate, if an
administrator wants to rotate it (there is no built-in for this). It
doesn't make sense to just delete a server certificate though, since
this will prevent other cluster members from connecting to the member
using that certificate. To delete a server certificate, the cluster
member should be removed entirely.

Signed-off-by: Mark Laing <mark.laing@canonical.com>