test/deps/import-busybox: stop adding `inittab`
Adding the `/etc/inittab` was a workaround added in 2015 but it seems to not be
required anymore. Also, the workaround used a leading `/` which annoyed `tar`
that needed to strip it to avoid path traversal attacks.

```
$ tar tvf c50b454b13f4f017b067e047e08ec53a19eb3ec9f7e4310ff9cb89a0b4a96127.tar.xz
...
drw-r--r-- 0/0               0 1969-12-31 19:00 rootfs/tmp/
drw-r--r-- 0/0               0 1969-12-31 19:00 templates/
-rw-r--r-- 0/0             236 1969-12-31 19:00 templates/template.tpl
-rw-r--r-- 0/0             380 1969-12-31 19:00 metadata.yaml
tar: Removing leading `/' from member names
-rw-r--r-- 0/0               1 1969-12-31 19:00 /rootfs/etc/inittab
```

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>

test/deps/import-busybox: use HTTPS link

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>