lxc: Prevent panic when overwriting the progress renderer
The progress renderer's UpdateOp gets added as a handler to the operation.
This allows the operation to update the progress on its own pace.

However when overwriting the initial progress renderer and assigning a new instance of it,
this also replaces the underlying mutex which can cause a race if the operation is in the process
of calling the original progress renderer's UpdateOp, obtains a lock of the mutex but the mutex
is in the process of getting replaced because a new instance of the progress renderer is assigned.

This causes the following error https://github.com/canonical/lxd-ci/actions/runs/15387812020/job/43290215796?pr=485.

In order to protect the initial mutex, assign the progress renderer for the actual export to a new variable.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 4f649a60ec6cd782a7d98e8a4c8c094c2afda384)