lxd/images: Support retrying image record deletion in autoUpdateImages
Previously the tx.DeleteImage error was not returned to the calling s.DB.Cluster.Transaction
function which meant the built in retry logic LXD has when the table is locked was not being
used. Instead return the error and let the retry mechanisms have a go at retrying the request.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>