lxd/db/db: Replaces EnterExclusive and ExitExclusive with RunExclusive
This wraps the exclusive access to the cluster database rather than relying on the
caller of EnterExclusive to call ExitExclusive with a transaction (even if not needed)
in order to release the lock.

This also avoids a bug with the old pattern whereby a panic could occur when an error
occurs due to double unlocking of the unlocked mutex.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>