test/README: drop `LXD_CONCURRENT` variable
This was introduced to skip 2 tests (`test_exec_concurrent` and
`test_concurrent`) deemed as very CPU intensive but in GH action runners which
have decent but not great performance they were pretty quick.

As of 2025-06, those 2 tests took this much time on GH action runners:

* 15s on `dir`
* 2m on `ceph`

The other drivers we all somewhere in between those 2 extremes.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>