test/main: reduce verbosity of `test-shell` closure
The `cleanup` handler must run when exiting a `test-shell` session but if the
last command returned non-0 (like `false`), we don't want to output the debug
information accompanying normal failures.

If a test script runs into an error, the `cleanup` handler will already have
reported the relevant debug info so there is no need to repeat it when exiting
the `test-shell` environment.

To do so, swallow any error code returned from the interactive \`test-shell\`.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>