lxd/storage: Add backupConfigCache struct with getters
Adds a new struct with the necessary getVolume and getPool getter funcs to allow temporarily
caching pools and volumes during the backup config creation.
When updating volumes which are used by many instances, this ensures either the pool or the volume
is only looked up from the DB and all subsequent get requests are served by cache.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>