lxd/storage: Add and implement GenerateInstanceCustomVolumeBackupConfig
Allow generating an instance's backup config for the used custom volumes.
Unlike the instance's main backup config, the backup config of the custom volumes might be included in multiple
instances.

To not require looking up all the entities from the DB for every instance, the custom volume's backup config
can be created independently from the actual instance's main backup config by specifying an optional cache.
The caller can decide to use this cache across multiple instances.
That is helpful in situations where a custom volume gets updated which causes the backup config files of all
instances using this volume to be updated.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>