lxd/auth/oidc: Update OIDC Verifier to use the secret func.
This additionally updates the OIDC Verifier to use v7 UUIDs as session
IDs. With this, we can determine when the session when created, thereby
allowing us to select the cluster secret that was used to encrypt the
cookies.

If the cluster secret that was used to encrypt OIDC credentials is out of
date, a new session is started with the latest secret.

Signed-off-by: Mark Laing <mark.laing@canonical.com>