lxd: Check network allocation used-by URLs.
Check the URLs returned by the network allocations endpoint.
Since most network types are shared from the default project (e.g.
bridge), IPs may be allocated to resources in many different projects.
This means that when getting allocations in a project that does not have
the network feature enabled, we need to get allocations from networks in
the default project, but we also need to ignore the effective project
when checking instance view access.

Closes #15855

Signed-off-by: Mark Laing <mark.laing@canonical.com>