test: split `DEBUG` into client, server and shell variables
This allows running with `LXD_DEBUG=server` to only pass the `--debug` flag
to the LXD daemon and not to every `lxc` client call. This is handy to avoid
the `lxc` echoing back the messages it received from `lxd`.

For backward compat, using `LXD_DEBUG=1` keeps passing `--debug` to `lxd` and
`lxc` as well as enabling shell tracing (`set -x`).

Similar handling is done for `LXD_VERBOSE`.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>