lxc/network: Allow listing allocations in non-default projects via project switch.
The `list-allocations` command had its own `--project` flag, separate to
the global `--project` override. This defaulted to "default" and was
always set as the project on the remote. This meant that using the
`--project` flag directly would always work, but if the current project
for the remote was changed, it would not be respected and the "default"
project would always be used.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit fc93e3ff8f47abb6d57c7556bea05120b7b2bfc1)