lxd/networks: Invoke network restore when requested
This reuses all of the pool loading mechanisms already in place for starting networks.
But it denies any retries as this is a costly operation already applied during start of the networks.
Also there wouldn't be any option to wait on this background operation as the network is already marked as being started.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 27b7d4bcd9ed6a0a8e74043bb568b4924361240e)